rewrote schema.sql, changed debate_page.html & flaskr.py
1) rewrote schema.sql: there are now 2 tables instead of 1. I changed
table “entries” to “debates”, and changed “id” to “did” (debate id,
will differentiate between aid, uid, etcetera). I added a sidea row and
a sideb row (these need formatting). I created table “arguments”, which
includes aid (int primary key autoincrement), did (integer), side
(integer), and argument (text).
2) arguments (debate_page.html): I copied the form from
createdebate.html and changed it to where you just type in your
argument and submit it. I put in a hidden field which contains the did
so that the arguments will only be shown in the debate they are written
in.
3) flaskr.py: I copied the add_entry() function and created a new route
with the function “add_argument()” it inserts the did and argument
which were passed from the form into the arguments table. I added onto
the debate_page(entries_did) function so that debate_page.html is
provided with the arguments from the query of the database.